Accounting for differences in the Fxns for .Net 5 object model
