+++ Consumer/EventGrid/Functions.cs
@@ -1,11 +1,14 @@
 using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.Azure.EventGrid.Models;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.EventGrid;
+
+#if NET5
+using Microsoft.Azure.Functions.Worker;
+#else
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json.Linq;
+#endif
 
 namespace Consumer.EventGrid
 {
@@ -17,51 +20,15 @@ namespace Consumer.EventGrid
         public static async Task EventGridProcessorAsync(
             [EventGridTrigger] EventGridEvent gridMessage,
             [EventHub(@"%CollectorEventHubName%", Connection = @"CollectorEventHubConnection")] IAsyncCollector<string> collector,
+#if !NET5
             ILogger log)
         {
-            var timestamp = DateTime.UtcNow;
-
-            var jsonMessage = JObject.FromObject(gridMessage);
-            var jsonContent = JObject.FromObject(gridMessage.Data);
-
-            var enqueuedTime = gridMessage.EventTime;
-            var elapsedTimeMs = (timestamp - enqueuedTime).TotalMilliseconds;
-
-            if (jsonContent.TryGetValue(@"workTime", out var workTime))
-            {
-                await Task.Delay(workTime.Value<int>());
-            }
-
-            var collectorItem = new CollectorMessage
-            {
-                MessageProcessedTime = DateTime.UtcNow,
-                TestRun = jsonContent.Value<string>(@"TestRunId"),
-                Trigger = @"EventGrid",
-                Properties = new Dictionary<string, object>
-                {
-                    { @"InstanceId", _instanceId },
-                    { @"ExecutionId", Guid.NewGuid().ToString() },
-                    { @"ElapsedTimeMs", elapsedTimeMs },
-                    { @"ClientEnqueueTimeUtc", enqueuedTime },
-                    { @"MessageId", jsonContent.Value<int>(@"MessageId") },
-                    { @"DequeuedTime", timestamp },
-                    { @"Language", @"csharp" },
-                }
-            };
-
-            await collector.AddAsync(collectorItem.ToString());
-
-            jsonMessage.Add(@"_elapsedTimeMs", elapsedTimeMs);
-            log.LogTrace($@"[{jsonContent.Value<string>(@"TestRunId")}]: Message received at {timestamp}: {jsonMessage}");
-
-            log.LogMetric("messageProcessTimeMs",
-                elapsedTimeMs,
-                new Dictionary<string, object> {
-                        { @"MessageId", jsonContent.Value<int>(@"MessageId") },
-                        { @"ClientEnqueuedTime", enqueuedTime },
-                        { @"DequeuedTime", timestamp },
-                        { @"Language", @"csharp" },
-                });
+#else
+            FunctionContext executionContext)
+        {
+            var log = executionContext.GetLogger(nameof(EventGridProcessorAsync));
+#endif
+            await ConsumerCode.EventGrid.ConsumeAsync(gridMessage, collector, log, _instanceId);
         }
     }
 }
