+++ Consumer/EventHubs/Functions.cs
@@ -1,11 +1,14 @@
 using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.Azure.EventHubs;
 using Microsoft.Azure.EventHubs.Processor;
 using Microsoft.Azure.WebJobs;
+
+#if NET5
+using Microsoft.Azure.Functions.Worker;
+#else
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json.Linq;
+#endif
 
 namespace Consumer.EventHubs
 {
@@ -18,61 +21,15 @@ namespace Consumer.EventHubs
             [EventHubTrigger(@"%EventHubName%", Connection = @"EventHubConnection", ConsumerGroup = "%EventHubConsumerGroupName%")] EventData[] ehMessages,
             PartitionContext partitionContext,
             [EventHub(@"%CollectorEventHubName%", Connection = @"CollectorEventHubConnection")] IAsyncCollector<string> collector,
+#if !NET5
             ILogger log)
         {
-            foreach (var ehMessage in ehMessages)
-            {
-                // replace 'body' property so output isn't ridiculous
-                var jsonMessage = JObject.FromObject(ehMessage);
-
-                var timestamp = DateTime.UtcNow;
-                var enqueuedTime = (DateTime)ehMessage.Properties[@"EnqueueTimeUtc"];
-                var elapsedTimeMs = (timestamp - enqueuedTime).TotalMilliseconds;
-
-                if (ehMessage.Properties.TryGetValue(@"workTime", out var value))
-                {
-                    await Task.Delay((int)value);
-                }
-
-                ehMessage.Properties.TryGetValue(@"TestRunId", out var testRunId);
-
-                var collectorItem = new CollectorMessage
-                {
-                    MessageProcessedTime = DateTime.UtcNow,
-                    TestRun = testRunId?.ToString(),
-                    Trigger = @"EventHub",
-                    Properties = new Dictionary<string, object>
-                    {
-                        { @"InstanceId", _instanceId },
-                        { @"ExecutionId", Guid.NewGuid().ToString() },
-                        { @"ElapsedTimeMs", elapsedTimeMs },
-                        { @"ClientEnqueueTimeUtc", enqueuedTime },
-                        { @"MessageId", ehMessage.Properties[@"MessageId"] },
-                        { @"DequeuedTime", timestamp },
-                        { @"Language", @"csharp" },
-                    }
-                };
-
-                await collector.AddAsync(collectorItem.ToString());
-
-                jsonMessage.Remove(@"Body");
-                jsonMessage.Add(@"Body", $@"{ehMessage.Body.Count} byte(s)");
-
-                jsonMessage.Add(@"_elapsedTimeMs", elapsedTimeMs);
-
-                log.LogTrace($@"[{testRunId?.ToString() ?? "null"}]: Message received at {timestamp}: {jsonMessage}");
-
-                log.LogMetric("messageProcessTimeMs",
-                    elapsedTimeMs,
-                    new Dictionary<string, object> {
-                        { @"PartitionId", partitionContext.PartitionId },
-                        { @"MessageId", ehMessage.Properties[@"MessageId"] },
-                        { @"SystemEnqueuedTime", enqueuedTime },
-                        { @"ClientEnqueuedTime", enqueuedTime },
-                        { @"DequeuedTime", timestamp },
-                        { @"Language", @"csharp" },
-                    });
-            }
+#else
+            FunctionContext executionContext)
+        {
+            var log = executionContext.GetLogger(nameof(EventHubProcessorAsync));
+#endif
+            await ConsumerCode.EventHubs.ConsumeAsync(ehMessages, partitionContext, collector, log, _instanceId);
         }
     }
 }
