+++ Consumer/ServiceBus/Functions.cs
@@ -1,11 +1,13 @@
 using System;
-using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Azure.ServiceBus;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json.Linq;
+
+#if NET5
+using Microsoft.Azure.Functions.Worker;
+#endif
 
 namespace Consumer.ServiceBus
 {
@@ -16,56 +18,32 @@ namespace Consumer.ServiceBus
         [FunctionName(nameof(ServiceBusQueueProcessorAsync))]
         public static async Task ServiceBusQueueProcessorAsync(
             [ServiceBusTrigger(@"%ServiceBusQueueName%", Connection = @"ServiceBusConnection", IsSessionsEnabled = true)] Message sbMessage,
-            [EventHub(@"%CollectorEventHubName%", Connection = @"CollectorEventHubConnection")]IAsyncCollector<string> collector,
+            [EventHub(@"%CollectorEventHubName%", Connection = @"CollectorEventHubConnection")] IAsyncCollector<string> collector,
+#if !NET5
             ILogger log)
         {
-            var timestamp = DateTime.UtcNow;
-            log.LogTrace($@"[{sbMessage.UserProperties[@"TestRunId"]}]: Message received at {timestamp}: {JObject.FromObject(sbMessage)}");
-
-            var enqueuedTime = sbMessage.ScheduledEnqueueTimeUtc;
-            var elapsedTimeMs = (timestamp - enqueuedTime).TotalMilliseconds;
-
-            if (sbMessage.UserProperties.TryGetValue(@"workTime", out var workTime))
-            {
-                await Task.Delay((int)workTime);
-            }
-
-            var collectorItem = new CollectorMessage
-            {
-                MessageProcessedTime = timestamp,
-                TestRun = sbMessage.UserProperties[@"TestRunId"].ToString(),
-                Trigger = @"ServiceBus",
-                Properties = new Dictionary<string, object>
-                {
-                    { @"InstanceId", _instanceId },
-                    { @"ExecutionId", Guid.NewGuid().ToString() },
-                    { @"ElapsedTimeMs", elapsedTimeMs },
-                    { @"ClientEnqueueTimeUtc", enqueuedTime },
-                    { @"SystemEnqueuedTime", enqueuedTime },
-                    { @"MessageId", sbMessage.MessageId },
-                    { @"DequeuedTime", timestamp },
-                    { @"Language", @"csharp" },
-                }
-            };
-
-            await collector.AddAsync(collectorItem.ToString());
-
-            log.LogMetric("messageProcessTimeMs",
-                elapsedTimeMs,
-                new Dictionary<string, object> {
-                    { @"Session", sbMessage.SessionId },
-                    { @"MessageNo", sbMessage.MessageId },
-                    { @"EnqueuedTime", enqueuedTime },
-                    { @"DequeuedTime", timestamp },
-                    { @"Language", @"csharp" },
-                });
+#else
+            FunctionContext executionContext)
+        {
+            var log = executionContext.GetLogger(nameof(ServiceBusQueueProcessorAsync));
+#endif
+            await ConsumerCode.ServiceBus.ConsumeAsync(sbMessage, collector, log, _instanceId);
         }
 
         [FunctionName(nameof(ClearDeadLetterServiceBusQueue))]
 #pragma warning disable IDE0060 // Remove unused parameter
-        public static void ClearDeadLetterServiceBusQueue([TimerTrigger("* 0 * * 1", RunOnStartup = true)]TimerInfo myTimer, ILogger log)
-#pragma warning restore IDE0060 // Remove unused parameter
+        public static void ClearDeadLetterServiceBusQueue([TimerTrigger("* 0 * * 1", RunOnStartup = true)] TimerInfo myTimer,
+#if !NET5
+            ILogger log)
         {
+#else
+            FunctionContext executionContext)
+        {
+            var log = executionContext.GetLogger(nameof(ServiceBusQueueProcessorAsync));
+#endif
+
+#pragma warning restore IDE0060 // Remove unused parameter
+
             var deadLetterQueueName = $@"{Environment.GetEnvironmentVariable("ServiceBusQueueName")}/$DeadLetterQueue";
             var client = new QueueClient(Environment.GetEnvironmentVariable(@"ServiceBusConnection"),
                     deadLetterQueueName, ReceiveMode.PeekLock);
