+++ Consumer/StorageQueues/Functions.cs
@@ -1,10 +1,12 @@
 using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.Storage.Queue;
+
+#if NET5
+using Microsoft.Azure.Functions.Worker;
+#else
 using Microsoft.Extensions.Logging;
-using Microsoft.WindowsAzure.Storage.Queue;
-using Newtonsoft.Json.Linq;
+#endif
 
 namespace Consumer.StorageQueues
 {
@@ -16,53 +18,15 @@ namespace Consumer.StorageQueues
         public static async System.Threading.Tasks.Task StorageQueueProcessorAsync(
             [QueueTrigger(@"%StorageQueueName%", Connection = @"StorageQueueConnection")] CloudQueueMessage queueMessage,
             [EventHub(@"%CollectorEventHubName%", Connection = @"CollectorEventHubConnection")]IAsyncCollector<string> collector,
+#if !NET5
             ILogger log)
         {
-            var timestamp = DateTime.UtcNow;
-
-            var jsonMessage = JObject.FromObject(queueMessage);
-            var jsonContent = JObject.Parse(queueMessage.AsString);
-
-            var enqueuedTime = jsonContent.Value<DateTime>(@"EnqueueTimeUtc");
-            var elapsedTimeMs = (timestamp - enqueuedTime).TotalMilliseconds;
-
-            if (jsonContent.TryGetValue(@"workTime", out var workTime))
-            {
-                await Task.Delay(workTime.Value<int>());
-            }
-
-            var collectorItem = new CollectorMessage
-            {
-                MessageProcessedTime = DateTime.UtcNow,
-                TestRun = jsonContent.Value<string>(@"TestRunId"),
-                Trigger = @"Queue",
-                Properties = new Dictionary<string, object>
-                {
-                    { @"InstanceId", _instanceId },
-                    { @"ExecutionId", Guid.NewGuid().ToString() },
-                    { @"ElapsedTimeMs", elapsedTimeMs },
-                    { @"ClientEnqueueTimeUtc", enqueuedTime },
-                    { @"SystemEnqueuedTime", queueMessage.InsertionTime },
-                    { @"MessageId", jsonContent.Value<int>(@"MessageId") },
-                    { @"DequeuedTime", timestamp },
-                    { @"Language", @"csharp" },
-                }
-            };
-
-            await collector.AddAsync(collectorItem.ToString());
-
-            jsonMessage.Add(@"_elapsedTimeMs", elapsedTimeMs);
-            log.LogTrace($@"[{jsonContent.Value<string>(@"TestRunId")}]: Message received at {timestamp}: {jsonMessage}");
-
-            log.LogMetric("messageProcessTimeMs",
-                elapsedTimeMs,
-                new Dictionary<string, object> {
-                        { @"MessageId", jsonContent.Value<int>(@"MessageId") },
-                        { @"SystemEnqueuedTime", queueMessage.InsertionTime},
-                        { @"ClientEnqueuedTime", enqueuedTime },
-                        { @"DequeuedTime", timestamp },
-                        { @"Language", @"csharp" },
-                });
+#else
+            FunctionContext executionContext)
+        {
+            var log = executionContext.GetLogger(nameof(StorageQueueProcessorAsync));
+#endif
+            await ConsumerCode.StorageQueues.ConsumeAsync(queueMessage, collector, log, _instanceId);
         }
     }
 }
