+++ Functions.sln
@@ -17,6 +17,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Consumer.net5", "Consumer.net5\Consumer.net5.csproj", "{26D824C1-D9DE-4BE4-B220-17999BAB233A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FunctionApp1", "FunctionApp1\FunctionApp1.csproj", "{0E40465E-BB7B-46B0-9F2B-C11673E3D413}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsumerCode", "ConsumerCode\ConsumerCode.csproj", "{AE9D8990-F6FE-41BD-94EE-976CD46073F5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -35,6 +39,14 @@ Global
 		{26D824C1-D9DE-4BE4-B220-17999BAB233A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{26D824C1-D9DE-4BE4-B220-17999BAB233A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{26D824C1-D9DE-4BE4-B220-17999BAB233A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0E40465E-BB7B-46B0-9F2B-C11673E3D413}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0E40465E-BB7B-46B0-9F2B-C11673E3D413}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0E40465E-BB7B-46B0-9F2B-C11673E3D413}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0E40465E-BB7B-46B0-9F2B-C11673E3D413}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AE9D8990-F6FE-41BD-94EE-976CD46073F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AE9D8990-F6FE-41BD-94EE-976CD46073F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AE9D8990-F6FE-41BD-94EE-976CD46073F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AE9D8990-F6FE-41BD-94EE-976CD46073F5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
